fix implicit float to int conversion

+++ src/Fgits/CarbonGermanHolidays/CarbonGermanHolidays.php
@@ -140,7 +140,7 @@ class CarbonGermanHolidays extends Carbon
             $states = array($states);
         }
 
-        $penanceDay = mktime(0, 0, 0, 11, 22 - ($year - 1 + $year / 4) % 7, $year);
+        $penanceDay = mktime(0, 0, 0, 11, 22 - ($year - 1 + ((int) ((int) ($year / 4)) % 7)), $year);
 
 
         // For all states
